requirements.txt
Flask==3.0.2
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography==42.0.5  # Required for JWT verification 